[SPARK-34212][SQL][FOLLOWUP] Move the added test to ParquetQuerySuite
This pr moves the added test from `SQLQuerySuite` to `ParquetQuerySuite`.

1. It can be tested by `ParquetV1QuerySuite` and `ParquetV2QuerySuite`.
2. Reduce the testing time of `SQLQuerySuite`(SQLQuerySuite ~ 3 min 17 sec, ParquetV1QuerySuite ~ 27 sec).

No.

Unit test.

Closes #32090 from wangyum/SPARK-34212.

Authored-by: Yuming Wang <yumwang@ebay.com>
